requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
langchain>=0.0.335
langchain-community>=0.0.10
langchain-experimental>=0.0.42
langchain-openai>=0.0.2
pydantic>=2.4.2
python-dotenv>=1.0.0
openai>=1.3.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1
